@@ -2,7 +2,7 @@ import 'package:dio/dio.dart';
 import 'package:stream_chat/src/core/error/error.dart';
 
 /// Error class specific to StreamChat and Dio
-class StreamChatDioError extends DioError {
+class StreamChatDioError extends DioException {
   /// Initialize a stream chat dio error
   StreamChatDioError({
     required this.error,
