@@ -88,7 +88,7 @@ void main() {
       requestOptions: options,
       data: errorResponse.toJson(),
     );
-    final err = DioError(requestOptions: options, response: response);
+    final err = DioException(requestOptions: options, response: response);
     final handler = ErrorInterceptorHandler();
 
     when(() => tokenManager.isStatic).thenReturn(false);
@@ -135,7 +135,7 @@ void main() {
         requestOptions: options,
         data: errorResponse.toJson(),
       );
-      final err = DioError(requestOptions: options, response: response);
+      final err = DioException(requestOptions: options, response: response);
       final handler = ErrorInterceptorHandler();
 
       when(() => tokenManager.isStatic).thenReturn(false);
@@ -153,7 +153,7 @@ void main() {
       } catch (e) {
         // need to cast it as the type is private in dio
         final error = (e as dynamic).data;
-        expect(error, isA<DioError>());
+        expect(error, isA<DioException>());
       }
 
       verify(() => tokenManager.isStatic).called(1);
@@ -179,7 +179,7 @@ void main() {
         requestOptions: options,
         data: errorResponse.toJson(),
       );
-      final err = DioError(requestOptions: options, response: response);
+      final err = DioException(requestOptions: options, response: response);
       final handler = ErrorInterceptorHandler();
 
       when(() => tokenManager.isStatic).thenReturn(true);
@@ -191,8 +191,8 @@ void main() {
       } catch (e) {
         // need to cast it as the type is private in dio
         final error = (e as dynamic).data;
-        expect(error, isA<DioError>());
-        final response = StreamChatNetworkError.fromDioError(error);
+        expect(error, isA<DioException>());
+        final response = StreamChatNetworkError.fromDioException(error);
         expect(response.errorCode, code);
       }
 
@@ -207,7 +207,7 @@ void main() {
       const path = 'test-request-path';
       final options = RequestOptions(path: path);
       final response = Response(requestOptions: options);
-      final err = DioError(requestOptions: options, response: response);
+      final err = DioException(requestOptions: options, response: response);
       final handler = ErrorInterceptorHandler();
 
       authInterceptor.onError(err, handler);
@@ -217,7 +217,7 @@ void main() {
       } catch (e) {
         // need to cast it as the type is private in dio
         final error = (e as dynamic).data;
-        expect(error, isA<DioError>());
+        expect(error, isA<DioException>());
       }
     },
   );
