@@ -21,7 +21,7 @@ void main() {
         statusCode: 200,
       );
 
-  DioError throwableError(
+  DioException throwableError(
     String path, {
     StreamChatNetworkError? error,
     bool streamChatDioError = false,
@@ -32,11 +32,11 @@ void main() {
       ..code = error?.code
       ..statusCode = error?.statusCode
       ..message = error?.message;
-    DioError? dioError;
+    DioException? dioError;
     if (streamChatDioError) {
       dioError = StreamChatDioError(error: error!, requestOptions: options);
     } else {
-      dioError = DioError(
+      dioError = DioException(
         error: error,
         requestOptions: options,
         response: Response(
@@ -210,7 +210,7 @@ void main() {
       await client.get(path);
     } catch (e) {
       expect(e, isA<StreamChatNetworkError>());
-      expect(e, StreamChatNetworkError.fromDioError(error));
+      expect(e, StreamChatNetworkError.fromDioException(error));
     }
 
     verify(() => dio.get(
@@ -263,7 +263,7 @@ void main() {
         await client.post(path);
       } catch (e) {
         expect(e, isA<StreamChatNetworkError>());
-        expect(e, StreamChatNetworkError.fromDioError(error));
+        expect(e, StreamChatNetworkError.fromDioException(error));
       }
 
       verify(() => dio.post(
@@ -317,7 +317,7 @@ void main() {
         await client.delete(path);
       } catch (e) {
         expect(e, isA<StreamChatNetworkError>());
-        expect(e, StreamChatNetworkError.fromDioError(error));
+        expect(e, StreamChatNetworkError.fromDioException(error));
       }
 
       verify(() => dio.delete(
@@ -371,7 +371,7 @@ void main() {
         await client.patch(path);
       } catch (e) {
         expect(e, isA<StreamChatNetworkError>());
-        expect(e, StreamChatNetworkError.fromDioError(error));
+        expect(e, StreamChatNetworkError.fromDioException(error));
       }
 
       verify(() => dio.patch(
@@ -425,7 +425,7 @@ void main() {
         await client.put(path);
       } catch (e) {
         expect(e, isA<StreamChatNetworkError>());
-        expect(e, StreamChatNetworkError.fromDioError(error));
+        expect(e, StreamChatNetworkError.fromDioException(error));
       }
 
       verify(() => dio.put(
@@ -486,7 +486,7 @@ void main() {
         await client.postFile(path, file);
       } catch (e) {
         expect(e, isA<StreamChatNetworkError>());
-        expect(e, StreamChatNetworkError.fromDioError(error));
+        expect(e, StreamChatNetworkError.fromDioException(error));
       }
 
       verify(() => dio.post(
