@@ -430,16 +430,15 @@ class StreamChannelState extends State<StreamChannel> {
       ],
       builder: (context, snapshot) {
         if (snapshot.hasError) {
-          var message = snapshot.error.toString();
-          if (snapshot.error is DioError) {
-            final dioError = snapshot.error as DioError?;
-            if (dioError?.type == DioErrorType.badResponse) {
-              message = dioError!.message ?? 'Bad response';
-            } else {
-              message = 'Check your connection and retry';
+          final error = snapshot.error;
+          if (error is DioException) {
+            if (error.type == DioExceptionType.badResponse) {
+              return Center(child: Text(error.message ?? 'Bad response'));
             }
+            return const Center(child: Text('Check your connection and retry'));
           }
-          return Center(child: Text(message));
+
+          return Center(child: Text(error.toString()));
         }
 
         final dataLoaded = snapshot.data?.every((it) => it) == true;
