@@ -1,12 +1,17 @@
 library stream_chat;
 
 export 'package:async/async.dart';
-export 'package:dio/src/cancel_token.dart';
-export 'package:dio/src/dio_error.dart';
-export 'package:dio/src/dio_mixin.dart' show Interceptor, InterceptorsWrapper;
-export 'package:dio/src/multipart_file.dart';
-export 'package:dio/src/options.dart';
-export 'package:dio/src/options.dart' show ProgressCallback;
+export 'package:dio/dio.dart'
+    show
+        DioException,
+        DioExceptionType,
+        RequestOptions,
+        CancelToken,
+        Interceptor,
+        InterceptorsWrapper,
+        MultipartFile,
+        Options,
+        ProgressCallback;
 export 'package:logging/logging.dart' show Logger, Level, LogRecord;
 export 'package:rate_limiter/rate_limiter.dart';
 export 'package:uuid/uuid.dart';
@@ -17,7 +22,6 @@ export 'src/client/key_stroke_handler.dart';
 export 'src/core/api/attachment_file_uploader.dart';
 export 'src/core/api/requests.dart';
 export 'src/core/api/responses.dart';
-export 'src/core/api/stream_chat_api.dart' show PushProvider;
 export 'src/core/api/stream_chat_api.dart';
 export 'src/core/error/error.dart';
 export 'src/core/http/interceptor/logging_interceptor.dart';
